+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+5.2.0 (2014-11-??)
+------------------
+
+  * added `stubbles\lang\Sequence`
+  * added `stubbles\streams\linesOf()`
+
+
 5.1.2 (2014-10-13)
 ------------------